pyproject.toml
[tool.poetry]
name = "api-usage-tracker"
version = "0.1.0"
description = "Django app to track API endpoint usage and mark deprecated endpoints"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2"
djangorestframework = "^3.14"
drf-spectacular = "^0.26"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3"
pytest-django = "^4.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
